# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Log Aggregator API
  description: |
    Высокопроизводительная система приёма, валидации, маршрутизации и хранения структурированных логов от микросервисов.
    
    ### Основные возможности
    - **Приём логов**: одиночный (`POST /api/v1/logs`) и пакетный (`POST /api/v1/logs/batch`).
    - **Пакетная обработка**: подробный отчет об обработке записей с флагом `retryable` для выборочной повторной отправки.
    - **Выборка и фильтрация**: фильтрация по сервису, уровню, подстроке, временному диапазону (RAM-буфер или файловый архив).
    - **Метрики и роутинг**: статистика в реальном времени (`/api/v1/stats`) и активные правила маршрутизации (`/api/v1/routes`).
    - **Healthcheck**: liveness (`/healthz`) и readiness (`/readyz`) пробы.
  version: 1.0.0
  contact:
    name: Log Aggregator Team
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки

tags:
  - name: Ingest
    description: Эндпоинты приёма логов от микросервисов
  - name: Logs
    description: Поиск, фильтрация и просмотр логов
  - name: Stats
    description: Метрики и системная статистика
  - name: Routing
    description: Правила маршрутизации логов
  - name: Probes
    description: Проверки работоспособности и готовности сервиса

paths:
  /api/v1/logs:
    post:
      tags:
        - Ingest
      summary: Приём одиночного лога
      description: |
        Принимает одну запись лога, выполняет валидацию обязательных полей (`service`, `level`, `message`),
        нормализует уровень логирования и помещает запись в неблокирующую очередь обработки.
        
        **Нормализация уровней:**
        - `trace`, `verbose`, `debug` → `debug`
        - `info`, `information` → `info`
        - `warn`, `warning` → `warn`
        - `err`, `error`, `fatal`, `critical`, `panic` → `error`
      operationId: ingestSingleLog
      requestBody:
        required: true
        description: Объект записи лога
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogEntry'
            example:
              service: auth-service
              level: error
              message: invalid credentials provided
              trace_id: req-98765-xyz
              host: auth-node-1
              fields:
                user_id: "1042"
                ip: "192.168.1.15"
      responses:
        '202':
          description: Запись успешно принята во внутреннюю очередь
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
        '400':
          description: Некорректный синтаксис JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "неверный JSON: unexpected end of JSON input"
        '422':
          description: Ошибка валидации структуры (отсутствуют обязательные поля или недопустимый уровень)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "отсутствует поле service"
        '503':
          description: Внутренняя очередь переполнена (backpressure). Клиенту рекомендуется повторить запрос позже
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "очередь переполнена, попробуйте позже"

    get:
      tags:
        - Logs
      summary: Поиск и фильтрация логов
      description: |
        Возвращает список записей логов, удовлетворяющих заданным критериям фильтрации.
        
        **Источники данных:**
        - `source=memory` (по умолчанию, если не указаны даты): быстрый кольцевой буфер в оперативной памяти.
        - `source=file` или `source=archive` (или при указании параметров `from`/`to`): дисковый архив JSON Lines файлов.
        
        **Заголовки ответа для пагинации:**
        - `X-Limit`: применённый размер страницы
        - `X-Offset`: текущее смещение
        - `X-Has-More`: `true`, если доступны следующие записи
      operationId: queryLogs
      parameters:
        - name: service
          in: query
          required: false
          description: Точное имя сервиса для фильтрации
          schema:
            type: string
          example: auth-service
        - name: level
          in: query
          required: false
          description: Уровень логирования
          schema:
            type: string
            enum: [debug, info, warn, error]
          example: error
        - name: search
          in: query
          required: false
          description: Подстрока для поиска внутри поля message (регистронезависимо)
          schema:
            type: string
          example: database connection
        - name: from
          in: query
          required: false
          description: Начало временного интервала (RFC3339 или YYYY-MM-DD[ HH:MM:SS])
          schema:
            type: string
          example: "2026-09-15T00:00:00Z"
        - name: to
          in: query
          required: false
          description: Конец временного интервала (RFC3339 или YYYY-MM-DD[ HH:MM:SS])
          schema:
            type: string
          example: "2026-09-15T23:59:59Z"
        - name: limit
          in: query
          required: false
          description: Максимальное количество записей в ответе (по умолчанию из server.default_query_limit, максимум server.max_query_limit)
          schema:
            type: integer
            minimum: 1
            default: 100
          example: 50
        - name: offset
          in: query
          required: false
          description: Количество пропускаемых записей от начала выборки
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: source
          in: query
          required: false
          description: Источник данных ("memory" — оперативная память, "file" / "archive" — дисковый архив)
          schema:
            type: string
            enum: [memory, file, archive]
          example: memory
      responses:
        '200':
          description: Массив найденных записей логов
          headers:
            X-Limit:
              schema:
                type: integer
              description: Применённый лимит записей
            X-Offset:
              schema:
                type: integer
              description: Смещение выборки
            X-Has-More:
              schema:
                type: boolean
              description: Флаг наличия следующих страниц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogEntry'
        '400':
          description: Некорректные параметры запроса (отрицательный offset, limit превышает максимум, неверный формат даты)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "параметр limit (5000) превышает допустимый максимум (1000)"
        '500':
          description: Ошибка чтения дискового архива
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/logs/batch:
    post:
      tags:
        - Ingest
      summary: Пакетный приём логов (Batch Ingest)
      description: |
        Принимает массив записей логов. Обрабатывает каждый элемент независимо, выполняя валидацию и постановку в очередь.
        
        ### Семантика кодов ответа и частичного приёма:
        - **`202 Accepted`**: Успешный приём (полный или частичный, `accepted > 0`). Если часть элементов была отклонена, они перечисляются в массиве `failed` с индексом, статусом и признаком `retryable`.
        - **`422 Unprocessable Entity`**: Полный отказ (`accepted == 0`), вызванный ошибками валидации записей (`retryable: false`). Клиент не должен повторять отправку без исправления записей.
        - **`503 Service Unavailable`**: Полный отказ (`accepted == 0`), вызванный исключительно переполнением очереди сервера (`retryable: true`). Клиент может повторить отправку батча позже с exponential backoff.
        - **`400 Bad Request`**: Синтаксическая ошибка JSON или передан пустой массив `logs`.
      operationId: ingestBatchLogs
      requestBody:
        required: true
        description: Пакет записей логов
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRequest'
            example:
              logs:
                - service: order-service
                  level: info
                  message: order #1001 created
                  trace_id: req-abc-1
                - service: order-service
                  level: warn
                  message: payment gateway response delayed
                  trace_id: req-abc-2
                - service: ""
                  level: error
                  message: broken record missing service name
      responses:
        '202':
          description: Пакет принят полностью или частично (accepted > 0)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
              examples:
                full_success:
                  summary: Полный успех (все записи приняты)
                  value:
                    accepted: 2
                    dropped: 0
                    errors: 0
                partial_success:
                  summary: Частичный успех (с деталями отклонённых записей)
                  value:
                    accepted: 1
                    dropped: 1
                    errors: 1
                    failed:
                      - index: 1
                        status: queue_full
                        error: "очередь переполнена, повторите попытку позже"
                        retryable: true
                      - index: 2
                        status: validation_error
                        error: "отсутствует поле service"
                        retryable: false
        '400':
          description: Неверный JSON или пустой массив логов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "батч пустой"
        '422':
          description: Полный отказ из-за ошибок валидации (accepted == 0 && errors > 0)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
              example:
                accepted: 0
                dropped: 0
                errors: 1
                failed:
                  - index: 0
                    status: validation_error
                    error: "отсутствует поле service"
                    retryable: false
        '503':
          description: Полный отказ из-за переполнения очереди сервера (accepted == 0 && dropped > 0)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
              example:
                accepted: 0
                dropped: 2
                errors: 0
                failed:
                  - index: 0
                    status: queue_full
                    error: "очередь переполнена, повторите попытку позже"
                    retryable: true
                  - index: 1
                    status: queue_full
                    error: "очередь переполнена, повторите попытку позже"
                    retryable: true

  /api/v1/stats:
    get:
      tags:
        - Stats
      summary: Метрики и системная статистика
      description: |
        Возвращает агрегированные счетчики:
        - Аптайм сервиса в читаемом виде
        - Число принятых, отброшенных и ошибочных запросов
        - Общее количество сохраненных логов и логов в горячем буфере RAM
        - Распределение логов по сервисам и уровням важности
      operationId: getStats
      responses:
        '200':
          description: Текущая статистика агрегатора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
              example:
                uptime: "5h 23m 15s"
                accepted: 1240
                dropped: 3
                errors: 1
                total_stored: 1240
                in_memory: 1240
                by_service:
                  auth-service: 410
                  order-service: 430
                  payment-service: 400
                by_level:
                  debug: 400
                  info: 600
                  warn: 160
                  error: 80

  /api/v1/routes:
    get:
      tags:
        - Routing
      summary: Список активных правил маршрутизации
      description: Возвращает список сконфигурированных правил fan-out маршрутизации логов.
      operationId: getRoutes
      responses:
        '200':
          description: Список активных правил маршрутизации
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteRule'
              example:
                - name: auth-errors
                  match:
                    service: auth-service
                    level: error
                  sinks:
                    - file
                - name: warn-and-above
                  match:
                    levels:
                      - warn
                      - error
                  sinks:
                    - stdout
                - name: default
                  match: {}
                  sinks:
                    - file
                    - stdout

  /healthz:
    get:
      tags:
        - Probes
      summary: Liveness-проба (проверка жизнеспособности процесса)
      description: Всегда возвращает HTTP 200, сигнализируя, что HTTP-сервер активен и обрабатывает запросы.
      operationId: checkHealth
      responses:
        '200':
          description: Процесс функционирует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok

  /readyz:
    get:
      tags:
        - Probes
      summary: Readiness-проба (проверка готовности принимать входящий трафик)
      description: |
        Проверяет готовность агрегатора к обслуживанию нагрузки:
        - Процесс не находится в состоянии graceful shutdown.
        - Инициализированы компоненты буфера памяти и хендлеров приёма.
      operationId: checkReady
      responses:
        '200':
          description: Сервер готов к приёму запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
              example:
                status: ready
        '503':
          description: Сервер не готов или находится в процессе остановки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
              example:
                status: not_ready
                reason: "сервер выключается или не готов к приёму запросов"

components:
  schemas:
    LogEntry:
      type: object
      required:
        - service
        - level
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения события на клиенте (опционально)
          example: "2026-09-15T21:00:00Z"
        server_timestamp:
          type: string
          format: date-time
          readOnly: true
          description: Время приёма и регистрации лога сервером агрегатора (заполняется автоматически)
          example: "2026-09-15T21:00:00.123456Z"
        service:
          type: string
          description: Идентификатор сервиса-источника
          example: auth-service
        level:
          type: string
          description: Уровень важности сообщения (debug, info, warn, error)
          enum: [debug, info, warn, error]
          example: error
        message:
          type: string
          description: Текст сообщения лога
          example: "database connection timeout after 3000ms"
        host:
          type: string
          description: Имя хоста или пода, сгенерировавшего лог
          example: auth-pod-7f98b
        trace_id:
          type: string
          description: Идентификатор распределённого трейса
          example: a1b2c3d4e5f6
        fields:
          type: object
          additionalProperties:
            type: string
          description: Дополнительные контекстные атрибуты ключ-значение
          example:
            user_id: "42"
            path: "/api/login"
            duration_ms: "3050"

    BatchRequest:
      type: object
      required:
        - logs
      properties:
        logs:
          type: array
          description: Непустой массив записей логов
          items:
            $ref: '#/components/schemas/LogEntry'

    BatchItemError:
      type: object
      required:
        - index
        - status
        - error
        - retryable
      properties:
        index:
          type: integer
          description: Порядковый номер (0-based) записи в исходном массиве logs
          example: 2
        status:
          type: string
          description: Категория сбоя (validation_error или queue_full)
          enum: [validation_error, queue_full]
          example: queue_full
        error:
          type: string
          description: Человекочитаемое описание причины отказа
          example: "очередь переполнена, повторите попытку позже"
        retryable:
          type: boolean
          description: true если отказ временный (queue_full) и клиент может повторить отправку без изменений
          example: true

    BatchResponse:
      type: object
      required:
        - accepted
        - dropped
        - errors
      properties:
        accepted:
          type: integer
          description: Количество успешно принятых в очередь записей
          example: 10
        dropped:
          type: integer
          description: Количество отброшенных записей из-за переполнения очереди
          example: 0
        errors:
          type: integer
          description: Количество записей с ошибками валидации структуры
          example: 0
        failed:
          type: array
          description: Список деталей по всем непринятым элементам (присутствует при dropped > 0 или errors > 0)
          items:
            $ref: '#/components/schemas/BatchItemError'

    StatsResponse:
      type: object
      required:
        - uptime
        - accepted
        - dropped
        - errors
        - total_stored
        - in_memory
        - by_service
        - by_level
      properties:
        uptime:
          type: string
          description: Время непрерывной работы процесса
          example: "5h 23m 15s"
        accepted:
          type: integer
          description: Суммарное число успешно принятых логов
          example: 1240
        dropped:
          type: integer
          description: Суммарное число отброшенных логов (переполнение очереди)
          example: 3
        errors:
          type: integer
          description: Суммарное число ошибок валидации/HTTP
          example: 1
        total_stored:
          type: integer
          description: Общее число сохранённых логов в хранилище
          example: 1240
        in_memory:
          type: integer
          description: Число логов в оперативном буфере
          example: 1240
        by_service:
          type: object
          additionalProperties:
            type: integer
          description: Распределение количества логов по сервисам
          example:
            auth-service: 410
            order-service: 430
            payment-service: 400
        by_level:
          type: object
          additionalProperties:
            type: integer
          description: Распределение количества логов по уровням
          example:
            debug: 400
            info: 600
            warn: 160
            error: 80

    RouteRule:
      type: object
      required:
        - name
        - match
        - sinks
      properties:
        name:
          type: string
          description: Уникальное имя правила
          example: auth-errors
        match:
          $ref: '#/components/schemas/MatchConfig'
        sinks:
          type: array
          items:
            type: string
          description: Список назначений для логов, подпадающих под правило
          example: [file, stdout]

    MatchConfig:
      type: object
      properties:
        service:
          type: string
          description: Фильтр по точному имени сервиса
          example: auth-service
        level:
          type: string
          description: Фильтр по одному уровню
          example: error
        levels:
          type: array
          items:
            type: string
          description: Фильтр по списку уровней
          example: [warn, error]
        message_regex:
          type: string
          description: Регулярное выражение для сопоставления с текстом сообщения
          example: "^auth:.*failed$"

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok

    ReadyResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ready
        reason:
          type: string
          description: Причина неготовности (присутствует при статусе not_ready)
          example: "сервер выключается или не готов к приёму запросов"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Описание возникшей ошибки
          example: "отсутствует поле service"